Modify ServiceClient instead of ServiceSimple
