+++ jqm-all/jqm-ws/pom.xml
@@ -360,6 +360,11 @@
             <artifactId>jqm-xml</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>com.enioka.jqm</groupId>
+            <artifactId>jqm-shared</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- Some log -->
         <dependency>
