+++ jqm-all/jqm-wstst/pom.xml
@@ -135,12 +135,6 @@
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>com.enioka.jqm</groupId>
-            <artifactId>jqm-ws</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
 
     </dependencies>
